refactor(frontend): 불필요한 파일 제거 및 디렉토리 생성
