+++ frontend/src/pages/App.js
@@ -1,5 +1,4 @@
-import logo from './logo.svg';
-import './App.css';
+import logo from '../assets/images/logo.svg';
 
 function App() {
   return (
